Refactor folder structure. Add Subtitle Manager class.
