+++ src/PostProductionScript/Models/ScriptModels/Script.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-using PostProductionScript.Interfaces;
+﻿using PostProductionScript.Interfaces;
 
 namespace PostProductionScript.Models.ScriptModels
 {
