+++ tests/PostProductionScript_Test/PostProductionScript_Test.csproj
@@ -25,4 +25,8 @@
     <ProjectReference Include="..\..\src\PostProductionScript\PostProductionScript.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Managers\" />
+  </ItemGroup>
+
 </Project>
